make note component
